+++ software_esp32/src/stmApp.cpp
@@ -291,7 +291,8 @@ void CStmApp::setMotorChars()
 {
     waitForFinishQueue=true;
     waitEEPFinished=true;
-    app_cmd(APP_PRE_SETMOTCHARS,String(motorChars.maxLowCurrent)+ARG_DELIMITER+String(motorChars.maxHighCurrent)+ARG_DELIMITER+String(motorChars.startOnPower)+ARG_DELIMITER+String(motorChars.noOfMinPulses));   
+    app_cmd(APP_PRE_SETMOTCHARS,String(motorChars.maxLowCurrent)+ARG_DELIMITER+String(motorChars.maxHighCurrent)+
+        ARG_DELIMITER+String(motorChars.startOnPower)+ARG_DELIMITER+String(motorChars.noOfMinCounts)+ARG_DELIMITER+String(motorChars.maxCalibRetries));   
     fastQueueMode=true;
 }
 
@@ -774,8 +775,11 @@ void  CStmApp::app_check_data()
                 motorChars.startOnPower=atoi(argptr[2]);
                 setupStartPosition(motorChars.startOnPower);
             }
-            if(argcnt == 4) {
-                motorChars.noOfMinPulses=atoi(argptr[3]);
+            if(argcnt >= 4) {
+                motorChars.noOfMinCounts=atoi(argptr[3]);
+            }
+            if(argcnt == 5) {
+                motorChars.maxCalibRetries=atoi(argptr[4]);
             }
             appState=APP_IDLE;
         }