+++ software_stm32/include/motor.h
@@ -95,6 +95,7 @@ int16_t appsetaction(char cmd, unsigned int valveindex, byte pos);
 extern uint8_t currentbound_low_fac;      // lower current limit factor for detection of end stop
 extern uint8_t currentbound_high_fac;     // upper current limit factor for detection of end stop
 extern uint8_t startOnPower;              // valve % on power start 
-extern uint16_t noOfMinPulses;
+extern uint16_t noOfMinCounts;
+extern uint8_t maxCalibRetries;
 
 #endif     //_MOTOR_H